在安装模块时正确设置目录和文件权限； 使用 /data/adb/modules_update 目录来存储新安装的模块，而非直接复制到 /data/adb/modules 目录，保证管理器能正确处理模块信息； 在安装时执行 customize.sh, 卸载时执行 uninstall.sh

+++ app/src/main/java/com/qimian233/ztool/utils/MagiskModuleManager.java
@@ -25,6 +25,10 @@ public class MagiskModuleManager {
 
         EnhancedShellExecutor executor = EnhancedShellExecutor.getInstance();
 
+        // 执行卸载脚本
+        EnhancedShellExecutor.ShellResult res0 = executor.executeRootCommand("sh " + MODULE_PATH + "/uninstall.sh");
+        if (!res0.isSuccess()) return context.getString(R.string.error_shell_command_failed, "Failed to execute uninstall.sh");
+
         // 删除模块目录
         EnhancedShellExecutor.ShellResult res1 = executor.executeRootCommand("rm -rf " + MODULE_PATH);
         if (!res1.isSuccess()) return context.getString(R.string.error_shell_command_failed, "Failed to remove module dir");
@@ -51,10 +55,18 @@ public class MagiskModuleManager {
             return context.getString(R.string.error_copy_assets);
         }
 
+        String MODULE_INSTALL_PATH = "/data/adb/modules_update/" + MODULE_ID;
+
         // 3. 使用 Root 移动到目标目录并设置权限
-        String cmd = "mkdir -p " + MODULE_PATH + " && " +
-                "cp -r " + tempDirPath + "/* " + MODULE_PATH + "/ && " +
-                "chmod -R 755 " + MODULE_PATH;
+        String cmd = "mkdir -p " + MODULE_INSTALL_PATH + " && " +
+                "cp -r " + tempDirPath + "/* " + MODULE_INSTALL_PATH + "/ && " +
+                "chmod -R 755 " + MODULE_INSTALL_PATH + " && " +
+                "chown -R 0:0 " + MODULE_INSTALL_PATH + " && " +
+                "chcon -R u:object_r:system_file:s0 " + MODULE_INSTALL_PATH + " && " +
+                "find " + MODULE_INSTALL_PATH + " -type d -exec chmod 755 {} \\; && " + // 目录权限 755
+                "find " + MODULE_INSTALL_PATH + " -type f -exec chmod 644 {} \\; && " + // 普通文件权限 644
+                "find " + MODULE_INSTALL_PATH + " -name '*.sh' -exec chmod 755 {} \\;" +
+                " && sh " + MODULE_INSTALL_PATH + "/customize.sh --install";
 
         EnhancedShellExecutor.ShellResult result = EnhancedShellExecutor.getInstance().executeRootCommand(cmd, 10);
